refactor: move function to handlePetAge.ts file
