+++ src/app/(app)/pets/components/CardPet.tsx
@@ -1,12 +1,14 @@
 'use client'
 
 import { Pet } from "@/types/PetsTypes";
-import { PetsContext } from '@/context/PetsContext';
 
+import Image from "next/image";
 import iconClose from '@/assets/svg/close.svg';
 import { IconArrow } from '@/components/icons/IconArrow';
-import Image from "next/image";
+
+import { PetsContext } from '@/context/PetsContext';
 import { useContext } from "react";
+import { handlePetAge } from "@/utils/handlePetAge";
 
 type CardPetType = {
   pet: Pet['View']
@@ -14,36 +16,6 @@ type CardPetType = {
 
 export function CardPet({ pet }: CardPetType) {
   const { submitDeletePet } = useContext(PetsContext)
-
-  function handleDateOfBirth(dateOfBirth: string | Date) {
-    const datePet = new Date(dateOfBirth);
-    const dateNow = new Date();
-  
-    const timeDiff = Math.abs(dateNow.getTime() - datePet.getTime());
-    
-    const dayDifference = Math.ceil(timeDiff / (1000 * 3600 * 24));
-
-    const years = Math.floor(dayDifference / 365);
-    const months = Math.floor((dayDifference % 365) / 30);    
-    const days = dayDifference % 30;
-  
-    let dateFormat: string = "";
-
-    if (years > 0) {
-      dateFormat = `${years} ${years > 1 ? 'anos' : 'ano'}`;
-      return dateFormat
-    }
-  
-    if (months > 0) {
-      dateFormat = `${months} ${months > 1 ? 'meses' : 'mês'}`;
-      return dateFormat
-    }
-
-    if (days > 0) {
-      dateFormat = `${days} ${days > 1 ? 'dias' : 'dia'}`;
-      return dateFormat
-    }
-  }
 
   return (
     <li 
@@ -63,7 +35,7 @@ export function CardPet({ pet }: CardPetType) {
         </span>
         
         <span className="text-xs text-[#2E2E2E] font-normal">
-          {handleDateOfBirth(pet.dateOfBirth)}
+          {handlePetAge(pet.dateOfBirth)}
         </span>
       </div>
